+++ poutyne/framework/model.py
@@ -225,18 +225,6 @@ class Model:
         Trains the network on a dataset. This method creates generators and calls
         the :func:`~Model.fit_generator()` method.
 
-        .. note:: With **Jupyter Notebooks**, a great number of displays per second (around > 200) seems to slow down
-            Jupyter Notebook. In which cases, we suggest to pass
-            ``progress_options={'show_every_n_train_steps': 100, 'show_on_valid': False}`` so that the number of
-            displays per second is at an acceptable level.
-
-        .. warning:: With **Jupyter Notebooks in Firefox**, if ``colorama`` is installed and colors are enabled (as it
-            is by default), a great number of epochs and steps per epoch can cause a spike in memory usage in Firefox.
-            The problem does not occur in Google Chrome/Chromium. To avoid this problem, you can decrease the number of
-            steps shown by passing ``progress_options={'show_every_n_train_steps': 100, 'show_on_valid': False}`` or you
-            can disable the colors by passing ``progress_options={'coloring': False}``. See
-            `this Github issue for details <https://github.com/jupyter/notebook/issues/5897>`__.
-
         Args:
             x (Union[~torch.Tensor, ~numpy.ndarray] or Union[tuple, list] of Union[~torch.Tensor, ~numpy.ndarray]):
                 Training dataset. Union[Tensor, ndarray] if the model has a single input.
@@ -348,18 +336,6 @@ class Model:
         Trains the network on a dataset. This method creates dataloaders and calls the
         :func:`~Model.fit_generator()` method.
 
-        .. note:: With **Jupyter Notebooks**, a great number of displays per second (around > 200) seems to slow down
-            Jupyter Notebook. In which cases, we suggest to pass
-            ``progress_options={'show_every_n_train_steps': 100, 'show_on_valid': False}`` so that the number of
-            displays per second is at an acceptable level.
-
-        .. warning:: With **Jupyter Notebooks in Firefox**, if ``colorama`` is installed and colors are enabled (as it
-            is by default), a great number of epochs and steps per epoch can cause a spike in memory usage in Firefox.
-            The problem does not occur in Google Chrome/Chromium. To avoid this problem, you can decrease the number of
-            steps shown by passing ``progress_options={'show_every_n_train_steps': 100, 'show_on_valid': False}`` or you
-            can disable the colors by passing ``progress_options={'coloring': False}``. See
-            `this Github issue for details <https://github.com/jupyter/notebook/issues/5897>`__.
-
         Args:
             train_dataset (~torch.utils.data.Dataset): Training dataset.
             valid_dataset (~torch.utils.data.Dataset): Validation dataset.
@@ -463,18 +439,6 @@ class Model:
         # pylint: disable=line-too-long
         """
         Trains the network on a dataset using a generator.
-
-        .. note:: With **Jupyter Notebooks**, a great number of displays per second (around > 200) seems to slow down
-            Jupyter Notebook. In which cases, we suggest to pass
-            ``progress_options={'show_every_n_train_steps': 100, 'show_on_valid': False}`` so that the number of
-            displays per second is at an acceptable level.
-
-        .. warning:: With **Jupyter Notebooks in Firefox**, if ``colorama`` is installed and colors are enabled (as it
-            is by default), a great number of epochs and steps per epoch can cause a spike in memory usage in Firefox.
-            The problem does not occur in Google Chrome/Chromium. To avoid this problem, you can decrease the number of
-            steps shown by passing ``progress_options={'show_every_n_train_steps': 100, 'show_on_valid': False}`` or you
-            can disable the colors by passing ``progress_options={'coloring': False}``. See
-            `this Github issue for details <https://github.com/jupyter/notebook/issues/5897>`__.
 
         Args:
             train_generator: Generator-like object for the training dataset. The generator must